LayoutTests/fast/forms/file/file-input-capture-expected.txt

Tests the capture attribute of <input type='file'>

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'capture' in input is true
PASS input.capture is ''
PASS input.capture is 'filesystem'
PASS input.capture is 'camera'
PASS input.capture is 'camcorder'
PASS input.capture is 'microphone'
PASS input.capture is 'filesystem'
PASS input.capture is 'camcorder'
PASS input.getAttribute('capture') is 'CamCorder'
PASS successfullyParsed is true

TEST COMPLETE
